Updated long values to int

+++ app/src/main/java/io/xxlabs/messenger/data/datatype/FactType.kt
@@ -1,12 +1,12 @@
 package io.xxlabs.messenger.data.datatype
 
-enum class FactType(val value: Long) {
+enum class FactType(val value: Int) {
     USERNAME(0),
     EMAIL(1),
     PHONE(2),
     NICKNAME(3);
 
     companion object  {
-        fun from(value: Long): FactType? = values().firstOrNull { it.value == value }
+        fun from(value: Int): FactType? = values().firstOrNull { it.value == value }
     }
 }
